requirements.txt
# this is the requirements.txt file...

#
# DEVELOPMENT / PRODUCTION
#

python-dotenv
pgeocode
pandas
requests
sendgrid==6.6.0
plotly.express 
requests
Flask
gunicorn
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-dateutil==2.8.1
six==1.16.0
SQLAlchemy==1.4.18
text-unidecode==1.3
Werkzeug==2.0.1

#
# TEST
#

pytest